requirements.txt
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
groq==0.11.0
python-dateutil==2.9.0.post0
rich==13.9.2
